Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> If we already tear down scratch in almost all cases, why not tear it down in every case?  We wouldn't need any of this margin in that case, right?
> 
> In other words, what cases does this margin help us with?  Flashing a same sized system.img without flashing a new super?  (And do we have the code to tear down just the system portion of the super in this case too? We'd want to do that even if we keep super around for its overlay of vendor or other partitions)

- Flashing a slightly increases system.img, vendor.img or (more likely) product.img without flashing a new super.
- We have code to tear down just the data associated with image being flashed.

[TL;DR]

You asked me to _document_ the reason we do this margin the code from a previous CL that added the margin, I am addressing that and increasing the margin b/c of increases in erase block sizes noted on device(s).

Yes, we have the code to tear down one partition's override data in fastbootd (only, not in bootloader fastboot, which causes another caveat on non-dynamic-partition devices).

We _only_ teardown scratch in the flashall case, or adb enable-verity. OTAs do not (yet). Flashing individual partitions does not, and will not, teardown the scratch partition (unless errors occur trying to mount and erase the data).

Rebuild, causing an increase of one (3 for good measure) erase block in an image even if due to normal sizing jitter. Flash just _one_ of the images (eg: vendor, product, system) and none of the others leaving their overrides in place. As noted in the comment and in previous CLs, this is to address the annoyance of having to flashall (to wipe scratch, which may not be intuitive given we are running out of dynamic partition space) in order to continue with development and (hopefully) quicker adb sync turnarounds.

This is all to balance between the obvious behaviors everyone expects and the caveats that happen when the scratch space is too large and prevents flashing individual partitions. This code is here to reduce the number of emails _I_ will receive asking me _why_ this is happening because no one reads the documentation.

OTAs remain broken when scratch is present. The CL b4 this pass dropping the relatively silent destruction of scratch solved that problem, we are still investigating how/when to call fs_mgr_overlayfs_teardown in ota engine.

Patch-set: 7
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>