liblp: Allocate "b" extents in reverse.
Allocate "b" extents backwards, so that A and B partitions wind up
clustered together. This will minimize fragmentation during A/B updates.

Note that this allocation strategy does not cause "b" partitions to be
allocated from the middle of "super", but rather, from the end. This
means that when a "b" partition grows, it will automatically fragment
(unlike an "a" partition which can be extended in-place). This is
unlikely to matter for two reasons. First, resizing a partition in-place
is already likely to fragment unless it was the most recent partition
added. Second, the OTA process does not resize in-place: it deletes all
partitions in the target slot, so allocation should be optimal.

Bug: 120433288
Test: liblp_test gtest
Change-Id: I1797e59e14c8b0d4d0e6855a1d984e8159b21df2