Make sure init symlink is added to INTERNAL_ROOT_FILES
$OUT/root/symlink provided by fake package isn't added into
INTERNAL_ROOT_FILES thus the symlink will not be created when
running specific targets like bootimage. This is fixed by adding
module path and stem to make up the target file name which will
be appended into INTERNAL_ROOT_FILES.

In the second part of the change custom build commands are used
to get rid of "-timestamp" module suffix which is added in
phony_package.mk.

Bug: 120460755
Test: make bootimage && ls -la $OUT/root/init
Change-Id: I0015332130a44322e06cecde8cd4cc56bc97c081