Update patch set 1
Patch Set 1:

(I'm looking for dwillemsen's review, since I don't have nearly enough grasp of this to do it right)

I was figuring that we would need two packages though, an init_system and an init_symlink or init_rootdir, since they're two different locations.  Does this still build correctly when building for only system, given that it's output target is TARGET_ROOT_OUT?  Is it possible that we drop the second_stage_init dependency in that case?  Or am I totally wrong here and this is sufficient?

On a side note, now that we mix TARGET_ROOT_OUT into system.img 100% of the time, can we just consolidate those build rules?  If part of this issue here is that we're maintaining two separate lists of files for root vs the rest of system, when we know that we're just going to combine them anyway, then long term we should merge those lists too?

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>