Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > Ehhh...  I should have said something on the bug.  I don't like the approach of just ignoring this.  Maybe we figure out that we're in a container and ignore this only in that case?  Or figure out how to get CAP_NET_ADMIN for ueventd in that case?
> > > 
> > > I don't want us to start ignoring this failure all of the time then get back into the situation where we had set 2MB but actually were getting less since setting SO_RCVBUF was capping us.
> > 
> > I guess then it depends on what's the desired behavior here. It seems like we don't want SO_RCVBUF to be capping us, so we are using RCVBUFFORCE to go beyond that cap. I am guessing we want some strong signal (guessing Logs isn't enough since they tend to get lost) that RCVBUF is capping us and RCVBUFFORCE isn't allowed? It would be a pretty strong signal though - uevent won't be able to start, and thus crashing the whole system eventually ...

That's pretty much exactly the strong signal that I want.  This is not going to change once a device has launched, so I want it to clearly crash during development.

And for processes other than ueventd, we get an SELinux audit when attempting to use RCVBUFFORCE when the permissions aren't there, so we'll see that in the SELinux boot tests (assuming it's a process that starts at boot).

> 
> Alternatively, maybe we can just try RCVBUFFORCE first always, and then fallback to RCVBUF if it fails? It's done similarly here: https://android-review.googlesource.com/c/platform/system/netd/+/427340
> 
> Although that wouldn't fix the concern of us getting capped by RCVBUF (if somehow RCVBUFFORCE never works correctly for some) since the fallback will be silent.

That's not going to work since as you recognized the fallback is silent.  But also, since processes that are requesting less than the cap won't need permissions for RCVBUFFORCE with the current code, but they'll always either need the permissions or throw an SELinux audit if they first try RCVBUFFORCE.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>