Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I've wanted to remove these for forever, but I've always been iffy since some devices in the ecosystem may need them and I didn't want to cause them churn.
> 
> This is made worse by Treble, which I think makes it a hard 'no' for removing them.  Unless we want to fork ueventd.rc into versions based on when the device was launched, aka ueventd.rc for devices <= P and this new one for devices > P.  Unless you disagree and think that this wouldn't be a treble violation.

Another way to deal with Treble constraints is to have partners maintain some parts of the old ueventd.rc that they need on new system.imgs to get system-only-ota to work. Provided that the delta (a subset of what's removed in this change) is small enough and that they actually do a system-only-ota, it shouldn't be too much work for partners.

Patch-set: 1
Reviewer: Gerrit User 1124693 <1124693@85c56323-6fa9-3386-8a01-6480fb634889>