Update patch set 2
Patch Set 2:

This command is *very* new, or so I thought, so if it does break scripts, I think it would be better to introduce a --no-wait option instead. The old UX of immediately printing "OK" when the operation may have failed is very confusing.

Patch-set: 2
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>