Do not remove speculative frames in all cases.
If the first frame of an unwind is a totally invalid pc that's not in any
map, a speculative frame is added. Rather than deleting this frame if no
more unwinding is possible, leave it. This fixes a case where the only
frame you get is an invalid one, but the speculative frame winds up in a shared
library or somewhere else and gets removed.

Bug: 120505086

Test: New unit tests to catch this case pass.
Test: Verified original crashing program now emits two backtrace lines.
Change-Id: I088dff21c057386dcdaeb3fc2578b24322683bd0