Revert "Set bin directories to 0751"
This reverts commit f4fc922f0b863659ca8e97c1f5fa522fafc7deb6.

Reason for revert: Chase banking app is crashing. This is because the
Chase app is using inotify_add_watch() on the /system/bin directory
and not checking the return value. The Android Security model
guarantees the immutability of files in /system/bin, so the inotify
watch is unnecessary. Until the Chase app fixes their bug, we need to
roll back this change.

Bug: 119605322
Change-Id: I02ea4f3898304aa8f359b44989f5a82d3dcfab86
Test: compiles