Update patch set 1
Patch Set 1:

Sounds reasonable to me.  This seems pretty reasonable too though, I'm not sure if that one is actually less complex.

I think the bigger question to me is does this need to be hardcoded to /sbin/poweroff?  We've been trying to move away from having a /sbin on Android, so if there's another option that would be ideal.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>