Update patch set 1
Patch Set 1:

Will Android itself make use of this functionality? How do we test it and prevent regressions? Is this something we want to take the burden of carrying in our tree if we're not using it?

In particular, no SELinux policy is written to support this functionality. Even if the kernel executed this binary, it wouldn't work due to missing policy. The code is essentially dead weight to us as currently implemented.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>