Update patch set 1
Patch Set 1:

Greg, do you think we might want something similar to this for cuttlefish in the long run?

The idea is that the VMM can tell the guest to shutdown, and wait for it, before it kills it.

Right now we just ctrl-c or otherwise explicitly kill the guest instance, I think.

Patch-set: 1
CC: Gerrit User 1062926 <1062926@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>