Update patch set 1
Patch Set 1:

While unrelated to this CL, the concept of the kernel executing userspace binaries is problematic for other reasons, such as the location of the files and environment variables (as pointed out by salyzyn), challenges with dealing with namespaces, and security policies. Userspace helpers are also used as an attack vector by exploit writers (eg https://googleprojectzero.blogspot.com/2018/09/oatmeal-on-universal-cereal-bus.html , section titled "From vold to init context"). 

A cleaner architecture would have been for the kernel to send a netlink message to userspace, where multiple processes could listen for the message and react accordingly. 

Again, this isn't related to this CL, but it does point to the fact that this may be a badly designed kernel API which we may not want to adopt.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>