Update patch set 1
Patch Set 1:

In essence, reboot executable is really only hiding the implementation detail of:

setprop sys.powerctl shutdown,now

which is underneath it all a UNIX domain socket connection to the Android init to start the shutdown process. Sort of the (netlink not) socket nnk@ refers to but talks to the appropriate process (init).

I am not stating this as an _alternative_ since Android properties are not to be used as an API (reboot command can be viewed as a HAL interface). But nnk@ does make a point I am getting even more concerned about regarding the poweroff= argument, reboot command has its own set of MAC (sepolicies) and DAC to execute and permit access to this property. If the kernel, through user helpers, runs the command, it needs to preserve this and _not_ run it as 'root' or it could be used as a means to elevating privileges as cited.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>