Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > Is it needed to extend the old fstab variant as well?
> 
> No, I think we are deprecating the old one, so there is no need.

I'd not add it for the old one in that case.  The less reason to use it the better.  Once we can get vold and the rest of the users moved over, we'll delete the old interface, so keeping it as is, seems best.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>