Update patch set 2
Patch Set 2:

(2 comments)

I understand the issue here, but I don't think this is the best way to solve it.  I don't think it's safe to try to force the other thread to shutdown and remove it from the LogTimesList in this fashion.

I've been reading up about safe ways to shut down a misbehaving thread like this.  It seems that the best option would be to use SO_SNDTIMEO to specify a timeout after which send() will return.  Once send() returns, we can then return back out to LogTimeEntry::threadStart() and let the thread shutdown naturally.  That will give us a bounded time that we would have to wait in this faulty situation.

I had also looked into what shutdown() can do, since that may be even better solution if it guarantees that the send() call would unblock and return with an error, but that doesn't seem to be guaranteed.

Lastly, we'd still need to handle not accepting more readers until the prune operation finishes, but we should be able to do that with the existing locks, right?  Otherwise, we'd just need an additional condition to prevent adding new readers to LastLogTimes.

Does that make sense?  I can try coding this myself soon.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>