Return only forwardings for chosen device in "adb -s $SERIAL forward --list".
Before the change, all the forwardings were returned even if
the client provided some serial number to identify a device (that serial was
ignored).  This change restricts output to forwardings active for chosen device,
if one was specified by a client (client-side code already provides a serial in
list command).

Change-Id: I870dd1257af4c59238372d44a910eb212580e5e7

Test: adb_test