Update patch set 3
Patch Set 3:

(3 comments)

> Patch Set 3:
> 
> > Patch Set 2:
> > 
> > (2 comments)
> > 
> > I understand the issue here, but I don't think this is the best way to solve it.  I don't think it's safe to try to force the other thread to shutdown and remove it from the LogTimesList in this fashion.
> > 
> > I've been reading up about safe ways to shut down a misbehaving thread like this.  It seems that the best option would be to use SO_SNDTIMEO to specify a timeout after which send() will return.  Once send() returns, we can then return back out to LogTimeEntry::threadStart() and let the thread shutdown naturally.  That will give us a bounded time that we would have to wait in this faulty situation.
> > 
> > I had also looked into what shutdown() can do, since that may be even better solution if it guarantees that the send() call would unblock and return with an error, but that doesn't seem to be guaranteed.
> > 
> > Lastly, we'd still need to handle not accepting more readers until the prune operation finishes, but we should be able to do that with the existing locks, right?  Otherwise, we'd just need an additional condition to prevent adding new readers to LastLogTimes.
> > 
> > Does that make sense?  I can try coding this myself soon.
> 
> Tom, Thank you very much for your valuable comments.
> (1) Actually, my previous solution is the "not accepting new readers when size-used too large", the same solution as you prefer. Now I am uploading it for you to review.  Hope it can satisfy/meet you high quality criteria.

Thank you; I do like this solution better.  The one addition in the other patchset that I think we need here too is to have maybePrune() retry prune() when prune() calls kickMe().  It seems that the race that a reader is kicked then reconnects is very easy to achieve, since maybePrune() won't be called until the next log message is received, which gives plenty of time for the reader to reconnect.

It seems that we could do this:
1) Set a flag in maybePrune() that we're in the pruning process (LogTimeEntry lock held to set the flag)
2) Call prune()
3) If prune() returns busy (or maybe we need an explicitly new flag if prune has called kickMe()), retry prune().
4) Unset the flag.

Note too that it would be important to release the LogTimeEntry lock each time that prune() returns in step 3), since LogTimeEntry::threadStart() will only be able to remove itself from LastLogTimes after it takes the lock.

Does something like that make sense?

> (2) shutdown DOES work.(My test has confirmed this). It seems to be "De facto standard"(in SO(stackoverflow) engineers:)).  We can use it.

Sounds good to me.  I messaged one of our internal network experts too, but they haven't gotten back to me.  If this is working successfully in practice for you, I'm okay moving forward with it.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>