Update patch set 1
Patch Set 1:

> Martin, I just want to get a confirmation that this shouldn't land until we remove the bionic libs from /system/lib. Is that right?

Isn't it the other way around? Without this we won't have the necessary namespace switching in place to be able to remove them.

Afaik this only blocks on the hwasan problem in b/121038155, which you got a WIP aosp/853909 for. I planned to wait for that, but another option is to go with the workaround patch eugenis@ suggested in http://b/119867084#comment7.

Patch-set: 1
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>