Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Martin, I just want to get a confirmation that this shouldn't land until we remove the bionic libs from /system/lib. Is that right?
> 
> Isn't it the other way around? Without this we won't have the necessary namespace switching in place to be able to remove them.
> 
> Afaik this only blocks on the hwasan problem in b/121038155, which you got a WIP aosp/853909 for. I planned to wait for that, but another option is to go with the workaround patch eugenis@ suggested in http://b/119867084#comment7.

Right, but if we land this now, libc.so will be double loaded in a process and it will cause problem like b/120661824. Could you run the test mentioned in b/120661824 with this to ensure that this won't break the test?

aosp/853909 needs to be refactored according to eugenis@. I suggest you to not wait for me. :)

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>