Update patch set 11
Patch Set 11: Presubmit-Ready+1

Rebased this patch onto the comically deep stack of string_view fixes to allow for using \0 instead of \1 as the delimiter, with the minor fixes needed to use it.

(The clang-format presubmit hook complains about this patch, because it's using an old, buggy version of clang-format that formats things badly. This is the result from switching the path from clang-stable to clang-r346389b)

Patch-set: 11
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1