Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > Debian seems to use /dev/disk/by-partlabel/ for what we call 'by-name', and has a slash-squashing mechanism for raw devices under /dev/disk/by-path/:
> > 
> > pci-0000:02:00.0-nvme-1
> > 
> > I guess the example given would be translated to "soc.0-f9824900.sdhci", or something like that.
> 
> It doesn't really help us to have two separate folders though, since the point here is that we're hardcoding checking /dev/block/by-name/ for the prefix of the super partition.
> 
> We already have /dev/block/platform/soc/624000.ufshc/sda10.  We could collapse that down into a single label with -'s instead of /'s, but that only helps if we also put those symlinks into /dev/block/by-name, which is a little ugly but maybe a reasonable idea if we don't want to get rid of 'by-name' from the prefix?
> 
> What would help cuttlefish the most here?

To directly unblock us with the least effort would just be to add everything into one namespace under by-name. Whether this is a symlink or not doesn't really matter.

This is being phrased as a cuttlefish issue, but IMO, Android has never before required any eMMC partition to have a specific name. The point of fstabs was to allow the user to map device names to mount points. This idea of a hard-coded "super" partition name that can't be probed for or searched by UUID as a requirement is the root cause of this problem.

So as long as we're saying the "super" hardcode can be reliably replaced in all places with the value of "androidboot.super_partition" and we just need that to appear under by-name, I'm OK with a solution that just symlinks exactly/only this named block device.

Patch-set: 4
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>