Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I'm not quite familiar with cuttlefish (cf)., but I guess this works because currently one device only has one partition in cf?

There are no partitions. It is a raw block device with no partition table. This is entirely permissible for Linux systems.

> Emulator has similar issue before but we had an intern to add the support for allowing multiple partitions in a single device. It seems done by modifying qemu to comply with by-name/uevent.partition_name though (Issac knows more details).

We might do this in the future for cuttlefish, but it's a big change, and for now it is not in scope. This is the first platform feature where having no partitions had any impact on functionality or conformance.

> https://docs.google.com/document/d/1fL7yVjYK9kqG-NAk4JAY0ujHhnDe3CWpeebh9sn3ywg/edit#heading=h.2qq1bb5kchqa
> 
> https://android-review.googlesource.com/c/platform/external/qemu/+/712100

Patch-set: 4
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>