Add the path to the bionic libs to the search paths
There will be two sets of bionic libs: the ones in the runtime APEX and
the ones in /system/bootstrap/lib[64]. The former is used for all
processes launched after the runtime APEX is activated and the latter is
used for a small number of processes (such as init, vold, and apexd
itself) that are started before the APEX is ready.

Since the dynamic linker has to use different bionic libs depending on
the context, the location of the bionic libs is abstracted behind the
symbol 'RUNTIME', which will be expanded to either
/system/bootstrap or /apex/com.androd.runtime.

The path to the bionic libs are added to the search paths of each linker
namespaces.

Bug: 120266448
Bug: 119274217
Test: m, device boots
Test: `lsof | grep libc` shows that libc is from /system/bootstrap/lib[64]

Change-Id: I409f79c1383686b85230422bdba3e2213fba20ee