Remove link from runtime to default in vendor proc
This commit removes the namespace fallback link from `runtime` to
`default` in vendor processes.  In the vendor process configuration, the
`default` namespace stands for `/vendor/${LIB}`.  Shared libraries in
com.android.runtime should be able to work without libraries under the
vendor partition (because it works in the system processes).

Test: n/a
Bug: 119867084
Change-Id: Ia06cbf9e6b644eaf04d7bbd06561c6617cb5bd3e