llkd: adbd shell setsid special case
A zombie process occurs when adb shell setsid <command> is issued,
however llkd can only detect if it is a result of a kernel livelock
by killing the associated parent, which would be adbd; resulting in
the adb connection(s) being terminated.  We will special case this
condition in order to preserve adbd for debugging purposes.

getComm() is effectively pure, so hold on to the return value for
sake of efficiency.

This also reverts commit 599958d1148c1137bee925f4cf08eda7f8050d98
which granted adbd blanket parent immunity from monitoring on
userdebug builds.  The new logic is a more refined means of
preserving the live lock checking associated with adbd.

Test: llk_unit_test
Test: date ; adb shell setsid sleep 900 ; date (reports 15 minutes)
Bug: 120983740
Change-Id: I6442463a48499d925a3a074423a24a1622905559