Support multi-fstype in fstab
The fs_type of system partition may not be ext4 (e.g., squashfs, erofs)
and we need to ensure a system partition of pure-android can be
mounted by init without modifications of fs_types in fstab.

Fstab example:
<src>  <mnt_point>  <type>  <mnt_flags and options> <fs_mgr_flags>
/dev/block/by-name/test /test ext4,f2fs  rw slotselect
"test" partition can be mounted successfully with a ext4 or f2fs in it.

Test: boot & check /proc/mounts
Change-Id: Ia94a292dff49ce07428badc6380ae00992307cb7