Don't use bind mount for /mnt/runtime/{write|read}/emulated
This is effectively a (probably temporary) revert of
fc592327a94dcfaac0f14b991ed7cab36c72551b

The read and write paths were created by bind-mounting the default path
and then re-mounting with appropriate sdcardfs options (gid and mask).

However, this is creating a problem when vold and other processes are
using different mount namespaces. Specifically, the read and write paths
are mounted with the correct options (gid = AID_EVERYBODY) in the mount
namespace of the vold process, but the same paths are mounted with
incorrect options (gid = AID_SDCARD_RW, which is the same as the default
path). It "looks like" the re-mount event are not correctly propagated
to other mount namespaces by sdcardfs.

Until we fix the root cause, revert it to unblock us.

Bug: 120266448
Test: device boots
Test: open camera, take picture -> picture is taken and can be seen in
the gallery app and the file app.

Change-Id: I1dc0427d6da0e90777e5b96073d44b69b6cc728e