Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> I still don't understand why this isn't refactored into a function that checks the verity state for a specific fstab_entry and then fs_mgr_update_verity_state() calls that function in a loop for the original callers.  Or, since there's only one other caller, just refactoring the whole thing and moving the looping logic to the caller, is probably better; it'll allow removing the callback too, since the function that takes a single fstab_entry can return everything directly.
> 
> fs_mgr_overlayfs.cpp, which is where I'm assuming the problems are stemming from, only uses this function to check for each fstab_entry that it checks if it has verity enabled already.  It already would be better off with this single function and the error reporting could be much more direct.

overlayfs collects them all (an early version checked for each entry, but that created even more noise and was the real reason we got all up in arms over the logspam), prior to looping through candidates so there is no reason for a refactor. AFAIC the logspam problem was solved when we collected them all just once.

The problem _today_ is that this error message creates FUD and engineering churn. The _best_ answer may be to not print anything at all ... except that someone may wish to know that (any) verity information is missing. So re-opening this CL is meant to discuss an appropriate solution so that we do not waste eng-man-days over the misinterpretation of this messaging.

Patch-set: 8
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0