Update patch set 1
Patch Set 1: Code-Review+1

LGTM.

What this means Yifan is that we haven't been registering the battery percentage when the device is suspended at all. healthd sets up the wake up alarm to check on the battery percentage every 10 mins without charger and every 1 min with charger. This is make sure the device can shut down gracefully if the battery percentage reaches close to or 0%.

The problem is, if the system is suspended, there is now way for healthd (and the framework as a result) to know that the battery is draining and is close to 0%. Thus resulting in abrupt shutdowns.

The reason it goes unnoticed is on a normal phone for example, I am guessing there is always something or the other waking up the device from suspend (network or applications setting up alarms for example). So, we never see the slack in the actual battery % vs what healthd knows.

I am only worried that this will add a periodic 10 mins wakeup alarm (which didn't exit before and we didn't see any problems with it yet).

Felix, if you can show and provide an example of how you stumbled upon this, it will be great, I will change this to +2 after we get that.

Patch-set: 1
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1