Set system property if 'updatable' process crashes too frequently
Critical processes currently cause the device to reboot into recovery when they crash 4 times in 4mins.
We extend this feature to a new 'updatable' class of init services. If these services crash 4 times in 4mins, instead of rebooting into recovery, we set ro.update_status to 'failure'. apexd and update_verifier will listen to that property and try to abort and rollback updates, staged apex modules and new slots respectively.

Test: Tested manually by marking zygote as updatable and killing zygote frequently, verified that property is set.
Bug: 120598832
Change-Id: I7d47ea1223f7792a834981c729694242ab3f28c9