Update patch set 1
Patch Set 1:

Hi, 

The 'updatable' keyword (it's not a class name) was introduced in order to delay the execution of the service until when the APEXes are activated. So, this is intended(meaningful) to the services that is started 'before' apexd is started in the *.rc files.

Since zygote is started well after apexd is started, we don't need to mark it as updatable, as there is no need delay the execution of zygote.

So, it looks like you are using the flag a little bit different from the purpose it was originally designed for. 

I assume that your goal is to mark processes that are directly/indirectly affected by the update of APEXes, right? If so, I would say that all processes that is started after apexd will be affected because they will all use bionic libs from the runtime APEX. It's not just zygote.

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>