Update patch set 1
Patch Set 1:

It may be worth figuring out why that parameter isn't being consumed by the kernel first then.  I'm still pretty open to this change, but I'd like to know if it's actually mandatory or if this is a kernel set up issue.  Ideally the current code works as is, and we get a clear error message if the arguments to the latter stages to init are incorrect, which this change would remove.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0