libvndksupport: Add android_is_in_vendor_context()
In preparation for cleaning up conditionally compiled VNDK libraries,
we need a function that determines if the current process is running in
the system context or the vendor context.

Also, use the version script for the library, so that libraries
depending on libvndksupport will be built the same, regardless of
whether a stub library of libvndksupport is used.  This is required for
libvndksupport to be used by no-vendor-variant VNDK libraries.

Bug: 119423884
Test: Add a dummy VNDK library that prints different string in system
      vs vendor context.  Use the library in different contexts and
      check the output.

Change-Id: Ic00927d8eadd2a21fcba9cb64cc93eaadd5bdd2d