Update patch set 1
Patch Set 1:

Considered that, one of the design considerations was keep this single threaded.

Creating a thread on demand would be one level worse than the sync() call. Creating a persistent thread is doable, but a semaphore to release the thread uses a futex which gets blocked by one of the classic livelocks when requiring a page under memory pressure. It just screams sync "is bad M'Kay" ...

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>