Update patch set 1
Patch Set 1:

liblog is _provisionally_ signal safe, we use trylocks and have alternate much less efficient paths that get triggered when the trylock fails. The reality is that we get log messages in signal handlers and after vfork and we accept the risk in the name of debugging.

I understand if the libbase logging, or to the kernel logs might not be safe.

This CL is more about 'how do I debug this' and less about 'what can permanently add to init to make this easier' ...

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>