Added filemap tool
This a tool to play FIEMAP I/O control to determine in various aspects
of how it works.

Dumps extents, reads file through block device, or bit inverts file.
The -c option is used to preallocate a file write and verify a pattern on it.
The -v option is used to verify a pattern written previously with -c.

Bug: 122702420
Test: ./filemap filemap
Test: ./filemap -r <block-device> <some-file> | cmp - <some-file>
Test: ./filemap filemap
Test: ./filemap -r <block-device> <file> | cmp - <file>
Test: ./filemap -c <block-device> <size> <file>
Test: ./filemap -v <block-device> <file>

Change-Id: If6b8906867ed8a6879d09cb35c793c6ca83be8e0
Signed-off-by: Ramon Pantin <pantin@google.com>